Extract BaseTaskRunProcessor
